requirements.txt
pymarc==3.0.3 
PyYAML==3.11
six==1.9.0

#added for rdf conversion
SPARQLWrapper==1.6.4
html5lib==0.999999
isodate==0.5.1
pyparsing==2.0.3
rdflib==4.2.0
rdflib-jsonld==0.3
requests==2.7.0

#for testing
nose==1.3.7

